a few refactors for bn254_blackbox_solver
